split mtl_start to three file with mtl_texture and mtl_buffer
